// packages/sero/types/models/guild-settings.types.ts
export enum GuildSettingType {
  ADMIN_ROLE = "admin-role",
  MODERATOR_ROLE = "moderator-role",

  WELCOME_CHANNEL = "welcome-channel",
  LEVEL_UP_CHANNEL = "level-up-channel",
  EXP_REWARD_DROP_CHANNEL = "exp-reward-drop-channel",

  BIRTHDAY_ROLE = "birthday-role",
  BIRTHDAY_CHANNEL = "birthday-channel",

  MEMBER_LOGS_CHANNEL = "member-logs-channel",
  MODERATION_LOGS_CHANNEL = "moderation-logs-channel",
  VC_LOGS_CHANNEL = "vc-logs-channel",
}